backend/requirements.txt
fastapi>=0.111
uvicorn>=0.29
httpx>=0.27
rembg>=2.0.57
opencv-python-headless>=4.9
Pillow>=10.3
numpy>=1.26
vtracer>=0.6
